--- CharlieBackend.Api.UnitTes/AutoData/Core/DefaultAutoDataAttribute.cs
@@ -1,16 +1,0 @@
-﻿using AutoFixture;
-using AutoFixture.Xunit2;
-using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace CharlieBackend.Api.UnitTest.AutoData.Core
-{
-    //public class DefaultAutoDataAttribute<TFixture> : AutoDataAttribute where TFixture : IFixture, new()
-    //{
-    //    public DefaultAutoDataAttribute() : base(new DefaultFixtureFactory<TFixture>().GetFixture)
-    //    {
-
-    //    }
-    //}
-}
