+++ CharlieBackend.Api.UnitTes/AutoData/CustomAutoDataAttributes/LessonsRepositoryAutoDataAttribute.cs
@@ -2,13 +2,12 @@
 using CharlieBackend.Api.UnitTest.AutoData.Core;
 using CharlieBackend.Api.UnitTest.AutoData.Fixtures;
 
-namespace CharlieBackend.Api.UnitTest.AutoData.CustomAutoDataAttribute
+namespace CharlieBackend.Api.UnitTest.AutoData.CustomAutoDataAttributes
 {
     public class LessonsRepositoryAutoDataAttribute : AutoDataAttribute
     {
         public LessonsRepositoryAutoDataAttribute() : base(new DefaultFixtureFactory<LessonsRepositoryFixture>())
         {
-
         }
     }
 }
